+++ packages/slate-plugins/src/components/Toolbar/Toolbar.tsx
@@ -16,7 +16,7 @@ export const Toolbar = styled.div<ToolbarProps>`
   box-sizing: content-box;
   user-select: none;
 
-  height: ${(props) => props.height || '18px'};
+  min-height: ${(props) => props.height || '18px'};
 
   > span,
   svg {