+++ packages/slate-plugins/src/elements/video/index.ts
@@ -1,4 +1,5 @@
 export * from './components';
+export * from 'elements/video/deserializeIframe';
 export * from './renderElementVideo';
 export * from './types';
 export * from './VideoPlugin';