+++ packages/slate-plugins/src/marks/highlight/deserializeHighlight.ts
@@ -2,8 +2,8 @@ import { DeserializeHtml } from 'common/types';
 import { getLeafDeserializer } from 'mark/utils';
 import { MARK_HIGHLIGHT } from './types';
 
-export const deserializeStrikethrough = (): DeserializeHtml => ({
-  leaf: getLeafDeserializer(MARK_HIGHLIGHT, {
-    tagNames: ['DEL', 'S'],
-  }),
+export const deserializeHighlight = ({
+  typeHighlight = MARK_HIGHLIGHT,
+}): DeserializeHtml => ({
+  leaf: getLeafDeserializer(typeHighlight),
 });