+++ packages/slate-plugins/src/marks/highlight/renderLeafHighlight.tsx
@@ -8,14 +8,15 @@ const HighlightText = styled.span<{ bg: string }>`
 `;
 
 export const renderLeafHighlight = ({
+  typeHighlight = MARK_HIGHLIGHT,
   bg = '#ffeeba',
 }: RenderLeafHighlightOptions = {}) => ({
   children,
   leaf,
 }: RenderLeafProps) => {
-  if (leaf[MARK_HIGHLIGHT])
+  if (leaf[typeHighlight])
     return (
-      <HighlightText data-slate-type={MARK_HIGHLIGHT} bg={bg}>
+      <HighlightText data-slate-type={typeHighlight} bg={bg}>
         {children}
       </HighlightText>
     );